add Ducks approach to organize the redux structure
