--- src/redux/reducers/reducersMainPage.js
@@ -1,15 +1,0 @@
-const initialState = {
-    posters: [],
-  };
-  
-  const mainPageReducer = (state = initialState, action) => {
-    switch (action.type) {
-      case 'SET_POSTERS':
-        return { ...state, posters: action.payload };
-      default:
-        return state;
-    }
-  };
-  
-  export default mainPageReducer;
-
