--- src/redux/reducers/reducersSessionList.js
@@ -1,14 +1,0 @@
-const initialState = {
-    films: [],
-  };
-  
-  const sessionListReducer = (state = initialState, action) => {
-    switch (action.type) {
-      case 'SET_FILMS':
-        return { ...state, films: action.payload };
-      default:
-        return state;
-    }
-  };
-  
-  export default sessionListReducer;
